requirements.txt
git+https://github.com/SYAN83/weibopy
pyyaml
pymongo
apscheduler